pyproject.toml
[project]
name = "atrypy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
]

[tool.ruff.lint]
select = ["ALL"]
